While the results from TdasPop are the same before and after this PR, the way the calculation is done has changed as described below:
- In both cases, the supplied value of `alpha_rate` was interpreted as
being alpha * (h/0.7)**3 in units of Num/Comoving Vol in Mpc^3/Yr in rest Frame
- The `PowerLawRates` class used to have a variable `alpha_rate` that
simply held the input value. The correct value of the rate was only
calculated in the function. We want to change this.

- A docstring was added to clarify the input value
- A test was introduced to check that the calculation was correct in a
limting case `test_limiting_case_easy` for this.
- Changed the `PowerLawRates` class to have a private attribute called
`_input_alpha_rate` that holds the input values (without any
multiplication).
- Changed the definition of `alpha_rate` attribute so that it is the
correct rate in the cosmology (includes whatever multiplication is
necessary for that).
- Fixed the rate evaluation method so that it uses this modified
alpha_rate variable and does not try to modify alpha_rate any more.
- Ran the test and checked that it passes.

	modified:   tests/test_ratedistribution.py
	modified:   tdaspop/rateDistributions.py
	modified:   tdaspop/version.py